Fix some properties
propSetValidBounds requires base <= top; however, this always holds if
setCapBounds is used in the way it is used in instructions.

Also strengthen the monotonicity property for setCapBounds: validBounds
is only required for exponent monotonicity, while the other properties,
in particular bounds monotonicity, don't require it.

Finally, fix a sealing property, as the return type of sealCap has
changed.