Add some more CRAM/CRRL properties
CRAM(CRRL(len)) == CRAM(len).  Also, if you have a request for bounds
that are a subset of the bounds of an existing input capability, then
aligning and rounding the requested bounds using CRAM/CRRL stays within
the original bounds of the input capability.

These properties don't hold for large lengths for which CRRL wraps
around to zero, however.